+++ driver-core/CHANGELOG.rst
@@ -33,6 +33,8 @@ CHANGELOG
   (JAVA-698)
 - [bug] Fix issue when executing a PreparedStatement from another cluster
   (JAVA-641)
+- [improvement] Allow Cluster subclasses to delegate to another instance
+  (JAVA-619)
 
 Merged from 2.0.9_fixes branch: